Refactor segment offset index to store only 4-byte positions for sequential indexes.
